Small improvements around Reason field and AlertCountTable draggable fields (#109239)
* Make count table field and alert table event renderer undraggable

* Remove tooltip from alert count table

* Remove DefaultDraggable wrapper from reason field

* Fix unit test and remove unused props

I am not 100% sure, but apparently, the test was broken due to a long import chain loop.
I cut the import chain by extracting some constants to a file.
