+++ x-pack/plugins/security_solution/public/timelines/components/row_renderers_browser/catalog/index.tsx
@@ -27,6 +27,7 @@ import {
   ThreatMatchExample,
   ZeekExample,
 } from '../examples';
+import { eventRendererNames } from './constants';
 import * as i18n from './translations';
 
 const Link = ({ children, url }: { children: React.ReactNode; url: string }) => (
@@ -39,26 +40,6 @@ const Link = ({ children, url }: { children: React.ReactNode; url: string }) => 
     {children}
   </EuiLink>
 );
-
-export const eventRendererNames: { [key in RowRendererId]: string } = {
-  [RowRendererId.alerts]: i18n.ALERTS_NAME,
-  [RowRendererId.auditd]: i18n.AUDITD_NAME,
-  [RowRendererId.auditd_file]: i18n.AUDITD_FILE_NAME,
-  [RowRendererId.library]: i18n.LIBRARY_NAME,
-  [RowRendererId.system_security_event]: i18n.AUTHENTICATION_NAME,
-  [RowRendererId.system_dns]: i18n.DNS_NAME,
-  [RowRendererId.netflow]: i18n.FLOW_NAME,
-  [RowRendererId.system]: i18n.SYSTEM_NAME,
-  [RowRendererId.system_endgame_process]: i18n.PROCESS,
-  [RowRendererId.registry]: i18n.REGISTRY_NAME,
-  [RowRendererId.system_fim]: i18n.FIM_NAME,
-  [RowRendererId.system_file]: i18n.FILE_NAME,
-  [RowRendererId.system_socket]: i18n.SOCKET_NAME,
-  [RowRendererId.suricata]: 'Suricata',
-  [RowRendererId.threat_match]: i18n.THREAT_MATCH_NAME,
-  [RowRendererId.zeek]: i18n.ZEEK_NAME,
-  [RowRendererId.plain]: '',
-};
 
 export interface RowRendererOption {
   id: RowRendererId;
