+++ x-pack/plugins/security_solution/public/timelines/components/timeline/body/renderers/reason_column_renderer.tsx
@@ -12,8 +12,7 @@ import React, { useCallback, useMemo, useState } from 'react';
 import styled from 'styled-components';
 import { BrowserFields, ColumnHeaderOptions, RowRenderer } from '../../../../../../common';
 import { Ecs } from '../../../../../../common/ecs';
-import { DefaultDraggable } from '../../../../../common/components/draggables';
-import { eventRendererNames } from '../../../row_renderers_browser/catalog';
+import { eventRendererNames } from '../../../row_renderers_browser/catalog/constants';
 import { ColumnRenderer } from './column_renderer';
 import { REASON_FIELD_NAME } from './constants';
 import { getRowRenderer } from './get_row_renderer';
@@ -53,12 +52,8 @@ export const reasonColumnRenderer: ColumnRenderer = {
       ? values.map((value, i) => (
           <ReasonCell
             key={`reason-column-renderer-value-${timelineId}-${columnName}-${eventId}-${field.id}-${value}-${i}`}
-            contextId={`reason-column-renderer-${timelineId}`}
             timelineId={timelineId}
-            eventId={eventId}
             value={value}
-            isDraggable={isDraggable}
-            fieldName={columnName}
             ecsData={ecsData}
             rowRenderers={rowRenderers}
             browserFields={browserFields}
@@ -81,26 +76,12 @@ const StyledEuiButtonEmpty = styled(EuiButtonEmpty)`
 `;
 
 const ReasonCell: React.FC<{
-  contextId: string;
-  eventId: string;
-  fieldName: string;
-  isDraggable?: boolean;
   value: string | number | undefined | null;
   timelineId: string;
   ecsData: Ecs;
   rowRenderers: RowRenderer[];
   browserFields: BrowserFields;
-}> = ({
-  ecsData,
-  rowRenderers,
-  browserFields,
-  timelineId,
-  value,
-  fieldName,
-  isDraggable,
-  contextId,
-  eventId,
-}) => {
+}> = ({ ecsData, rowRenderers, browserFields, timelineId, value }) => {
   const [isOpen, setIsOpen] = useState(false);
 
   const rowRenderer = useMemo(() => getRowRenderer(ecsData, rowRenderers), [ecsData, rowRenderers]);
@@ -111,7 +92,7 @@ const ReasonCell: React.FC<{
       rowRenderer.renderRow({
         browserFields,
         data: ecsData,
-        isDraggable: true,
+        isDraggable: false,
         timelineId,
       })
     );
@@ -136,29 +117,21 @@ const ReasonCell: React.FC<{
 
   return (
     <>
-      <DefaultDraggable
-        field={fieldName}
-        id={`reason-column-draggable-${contextId}-${eventId}-${fieldName}-${value}`}
-        isDraggable={isDraggable}
-        value={`${value}`}
-        tooltipContent={value}
-      >
-        {rowRenderer && rowRender ? (
-          <EuiPopover
-            isOpen={isOpen}
-            anchorPosition="rightCenter"
-            closePopover={handleClosePopOver}
-            button={button}
-          >
-            <EuiPopoverTitle paddingSize="s">
-              {i18n.EVENT_RENDERER_POPOVER_TITLE(eventRendererNames[rowRenderer.id] ?? '')}
-            </EuiPopoverTitle>
-            {rowRender}
-          </EuiPopover>
-        ) : (
-          value
-        )}
-      </DefaultDraggable>
+      {rowRenderer && rowRender ? (
+        <EuiPopover
+          isOpen={isOpen}
+          anchorPosition="rightCenter"
+          closePopover={handleClosePopOver}
+          button={button}
+        >
+          <EuiPopoverTitle paddingSize="s">
+            {i18n.EVENT_RENDERER_POPOVER_TITLE(eventRendererNames[rowRenderer.id] ?? '')}
+          </EuiPopoverTitle>
+          {rowRender}
+        </EuiPopover>
+      ) : (
+        value
+      )}
     </>
   );
 };
